+++ lib/data/data_structures/Runway/Runway.h
@@ -8,12 +8,12 @@ public:
          const Vector3 &endPositionMeters); // Start position, end position
 
   Runway(std::string name, const Vector3 &startPositionMeters,
-         double distanceMeters,
+         float distanceMeters,
          const Vector3 &direction); // Start position, distance, direction
   Vector3 getStartPointMeters();
   Vector3 getEndPointMeters();
   Vector3 getDirection();
-  double getLengthMeters();
+  float getLengthMeters();
 
   std::string toJSON();
 
@@ -23,6 +23,6 @@ private:
   Vector3 startPositionMeters;
   Vector3 endPositionMeters;
   Vector3 direction;
-  double lengthMeters;
+  float lengthMeters;
   std::string name;
 };
