+++ lib/global/math/PID/PID.cpp
@@ -4,7 +4,7 @@
 pid::pid() : kP(0), kI(0), kD(0), error(0), now(0), prevError(0), prevTime(0), integral(0), derivative(0) {}
 
 // WARNING: NOT FINISHED!
-pid::pid(double KP, double KI, double KD) {
+pid::pid(float KP, float KI, float KD) {
     this->kP = KP;
     this->kI = KI;
     this->kD = KD;
@@ -14,10 +14,10 @@ pid::pid(double KP, double KI, double KD) {
 }
 
 // WARNING: NOT FINISHED!
-double pid::compute(double input, double target) {
-    double error = target - input;
+float pid::compute(float input, float target) {
+    float error = target - input;
     unsigned long now = 0; // HAL_GetTick();
-    double dt = (now - this->prevTime) / 1000.0f;
+    float dt = (now - this->prevTime) / 1000.0f;
     if (dt >= 0.001f) {
         this->derivative = (error - this->prevError) / dt;
     }
